Add list-like .remove() method
Allow removing objects from lists by object
identity rather than list index. We can't use
the current remove_foo(x) method for this as that
throws the wrong type of exception if x isn't in
the list (and none at all in debug mode) so add
a similar Python-specific method for this purpose.