Fix ambiguous hash_value with Boost 1.81
We have a templated hash_value function for any
class T but as of Boost 1.81 (which includes a
major overhaul of boost.container_hash) there is
a similar function for range types in Boost.
Add more specific overloads or templates for
some IMP classes so the compiler can disambiguate
calls to hash_value.