Do our own handling of polymorphic pointers
In order to correctly serialize a polymorphic pointer
we need the most-derived type. cereal includes machinery
for this but it relies on the linker to make sure that
certain objects are unique in the process, and this doesn't
work well with link time optimization, or on Windows,
as per USCiLab/cereal#783. Provide our own similar
mechanism that registers Object subclasses in precisely
one place - the Object class itself in IMP.kernel.