Store temporary ScoreStates in set, not vector
When we have a large number of Restraints in the system,
each of which references one of a small number of ScoreStates,
calculating the required ScoreStates is very expensive
because the temporary vector gets many copies of each ScoreState
pointer. Eliminate this by using a set instead. This has
a higher insertion cost (since the entries are sorted)
but we end up having to sort the list to remove
duplicates anyway (and can now skip that step).