Don't duplicate Objects during (de-)serialization
Add pointer tracking to serialization of IMP::Pointer
and friends. Each pointed-to object is assigned a
unique ID; if we see the same ID again, we can skip
serializing the object (and so avoid creating a copy
of it on deserialization). This (ab)uses cereal's
existing infrastructure for tracking of std::shared_ptr.