Add sequence offset to Chain decorator
To support models where residue indexes do not
start at 1, add an integer offset to the Chain
decorator, which defaults to 0. This behaves
similarly to python-ihm's auth_seq_id_map, which
maps seq_ids (which always start at 1) to
author-provided residue numbering.

This is currently handled in an ad hoc fashion
in PMI, by adding gaps to the start of the FASTA
sequence, but that can only support positive
offsets.