PDBSelector now works with PDBRecord, not raw line
Rather than passing a raw line of text from the PDB
file to PDBSelector, pass a PDBRecord type. This wraps
either the PDB text or the equivalent parsed data
from the atom_site table in an mmCIF file, and provides
methods to get atom/residue name, etc. This ensures
that when working with mmCIF we are not restricted
to PDB format limitations, such as single-character
chain IDs or insertion codes. Closes #1083.