Add compiler info to IMP CMake config
Add information on the C++ compiler that was used
to build IMP itself, plus any needed flags, to the
generated CMake config file. This can be used to
build binaries that link against IMP using the same
compiler flags, and we can also use it to run our
compilation tests; this is more reliable than trying
to parse the CMake cache, since not all information is
in there.