Try to suppress linkage against C++ MPI bindings
The MPI C++ bindings have long since been deprecated,
and we don't use them in IMP, so try to prevent linkage
against them. For example, building IMP on RHEL8 with
OpenMPI currently links against libmpi_cxx.so.40, which
then fails on a Fedora 40 box as the C++ bindings have
been removed in OpenMPI 5.