Don't try to suppress use of MPI C++ binding
This doesn't really work because we don't currently
propagate the necessary compile flags, and CMake tries
to link against the mpi_cxx library anyway.
Instead, we'll set the necessary config globally in our
build system to build RHEL packages that also work
on Fedora 40 or later (where the C++ bindings are gone).

This reverts commit 7bfed8c07ca5e8ea5a4c74958182a5c74936a42e.
This reverts commit 841a81d3fbbe255b97b8a338faa31e4a2bc67a17.