Limit memory usage in get_solvent_accessibility()
If given a very large protein, FoXS can segfault as
the grid used to compute solvent accessibility is so
large that its number of voxels overflows an int.
Avoid this by increasing the grid spacing if necessary
to cap memory usage to ~20GiB. This is currently hard
coded but could be passed as a parameter in future.