Add patch to build with NumPy 1/2 on RHEL 8
The copy of the NumPy 2 headers, which adds
support for both NumPy 1 and NumPy 2 in our
Ubuntu builds, does not work in RHEL 8 because
they build for NumPy 1.16 or later, and RHEL 8
ships with NumPy 1.14. Declare an older NumPy
feature version and look for the API info in
the older location if necessary. This patch is
used in the imp_nightly_builds repo to build
binaries on RHEL8 that should also work in Fedora.