Allow marking a test as 'unstable'
'unstable' tests are tests that do not reliably
pass or fail, such as 'science' tests that perform
some sort of stochastic sampling or optimization
and then assert on the results. The new IMP.test.unstable
decorator can be used to mark such tests. They are then
skipped if the IMP_SKIP_UNSTABLE_TESTS environment
variable is set. We will use this to make our CI
reliably pass or fail by skipping unstable tests.