Include ABI tag in Python modules if asked
We currently generate Python extension modules
with a plain .so extension (on Mac or Linux).
This doesn't include Python version information
so could cause a crash if we try to use them from
a different version of Python. Add a new CMake
flag, IMP_USE_PYTHON_SOABI, which (if turned on)
will add this information, generating extension
modules with names like
_IMP_kernel.cpython-313-x86_64-linux-gnu.so
instead of
_IMP_kernel.so